Updated node selection and filtering.
With this update, nodes that belong to residues with at least one node in contact will be kept in the system.
This prevents issues where nodes are removed for lack of contacts but their representing atoms are kept for being
part of a residue in contact with other residues.
Using the notSameRes=False option in dnap.filterContacts will keep the method from removing edges between nodes in
the same residue. This may seem like a similar results as the code that was introduced here, but it is not since this
commit will allow edges between nodes in the same residue to be removed while keeping the nodes themselves in the system.